Conflicts:
	dao/src/main/java/com/aplana/sbrf/taxaccounting/dao/impl/ReportPeriodDaoImpl.java
	migration-ejb/src/main/java/com/aplana/sbrf/taxaccounting/service/MessageServiceBean.java
	service/src/main/java/com/aplana/sbrf/taxaccounting/service/impl/DeclarationDataServiceImpl.java

+++ gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/module/refbookdata/client/editform/AbstractEditView.java
@@ -319,7 +319,7 @@ public abstract class AbstractEditView extends ViewWithUiHandlers<EditFormUiHand
         } else {
             if (value.getValue() != null && !value.getValue().toString().trim().isEmpty()) {
                 String valStr = (String) value.getValue();
-                int val = Integer.valueOf(valStr);
+                double val = Double.valueOf(valStr);
                 if (val < 0){
                     throw new NumberFormatException();
                 }
